% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{append_deps}
\alias{append_deps}
\title{Append dependencies}
\usage{
append_deps(filename)
}
\arguments{
\item{filename}{The name of the file containing the dependency definitions.}
}
\description{
Append further dependencies to the list of available dependencies, so that they can
  be used by \code{\link{r2deck}} or \code{\link{r2mapbox}}.
}
