% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{carto_style}
\alias{carto_style}
\title{Carto style}
\usage{
carto_style(theme = "dark-matter")
}
\arguments{
\item{theme}{The theme of the style, \code{dark-matter}, \code{positron} or \code{voyager}.}
}
\description{
Carto style
}
