% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{deck_properties}
\alias{deck_properties}
\title{Deck properties}
\usage{
deck_properties(map_style = carto_style("dark-matter"),
  longitude = -122.45, latitude = 37.8, zoom = 10, ...)
}
\arguments{
\item{map_style}{The style definition of the map conforming to the Mapbox Style Specification.}

\item{longitude}{The longitude of the initial geographical center point of the map.}

\item{latitude}{The latitude of the initial geographical center point of the map.}

\item{zoom}{The initial zoom level of the map.}

\item{...}{Further options that are passed to the Map class.}
}
\description{
Deck properties
}
