% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{make_mapbox_properties}
\alias{make_mapbox_properties}
\title{Mapbox properties}
\usage{
make_mapbox_properties(style = get_carto_style("dark-matter"),
  center = c(-122.45, 37.8), zoom = 6, ...)
}
\arguments{
\item{style}{The style definition of the map conforming to the Mapbox Style Specification.}

\item{center}{The initial geographical center point of the map.}

\item{zoom}{The initial zoom level of the map.}

\item{...}{Further options that are passed to the Map class.}
}
\description{
Mapbox properties
}
