% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{make_r2deck_viz}
\alias{make_r2deck_viz}
\title{Render r2deck using header def in JavaScript file}
\usage{
make_r2deck_viz(script = NULL)
}
\arguments{
\item{script}{The name of the JavaScript file containing the vizualisation function.}
}
\description{
Render r2deck using header def in JavaScript file
}
