% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{mapbox_properties}
\alias{mapbox_properties}
\title{Mapbox properties}
\usage{
mapbox_properties(center = c(-122.45, 37.8), zoom = 6,
  style = "https://basemaps.cartocdn.com/gl/dark-matter-gl-style/style.json",
  ...)
}
\arguments{
\item{center}{The initial geographical center point of the map.}

\item{zoom}{The initial zoom level of the map}

\item{style}{The style definition of the map conforming to the Mapbox Style Specification.}

\item{...}{Further options that are passed to the Map class.}
}
\description{
Mapbox properties
}
