% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{r2deck-shiny}
\alias{r2deck-shiny}
\alias{r2deck_output}
\alias{render_r2deck}
\title{Shiny bindings}
\usage{
r2deck_output(output_id, width = "100\%", height = "400px")

render_r2deck(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{output_id}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a r2deck}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Shiny bindings
}
