% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{r2deck}
\alias{r2deck}
\title{Create a widget object}
\usage{
r2deck(script, data, lng = -122.45, lat = 37.8, zoom = 8,
  map_style = get_carto_style(), options = NULL, width = NULL,
  height = NULL, element_id = NULL, ...)
}
\arguments{
\item{script}{The name of the JavaScript file containing the vizualisation function.}

\item{data}{The data that is passed to the vizualisation function.}

\item{lng}{The longitude of the initial geographical center point of the map.}

\item{lat}{The latitude of the initial geographical center point of the map.}

\item{zoom}{The initial zoom level of the map.}

\item{map_style}{The style definition of the map conforming to the Mapbox Style Specification.}

\item{options}{(optional) Additional options that are passed to the vizualisation function as parameter \code{options}.}

\item{width}{(optional) The width of the widget.}

\item{height}{(optional) The height of the widget.}

\item{element_id}{(optional) The unique id of the widget.}

\item{...}{Additional properties (picking radius, ...) that are passed to the Deck class.}
}
\description{
Create a widget object
}
