% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapbox.R
\name{r2mapbox}
\alias{r2mapbox}
\title{Create a widget object using the Mapbox WebGL context}
\usage{
r2mapbox(script, data, center = c(-122.45, 37.8), zoom = 8,
  style = get_carto_style("dark-matter"), options = NULL,
  width = NULL, height = NULL, element_id = NULL, ...)
}
\arguments{
\item{script}{The name of the JavaScript file containing the vizualisation function.}

\item{data}{The data that is passed to the vizualisation function.}

\item{center}{...}

\item{zoom}{...}

\item{style}{The style definition of the map conforming to the Mapbox Style Specification.}

\item{options}{(optional) Additional options that are passed to the vizualisation function as parameter \code{options}.}

\item{width}{(optional) The width of the widget.}

\item{height}{(optional) The height of the widget.}

\item{element_id}{(optional) The unique id of the widget.}

\item{...}{Additional properties (picking radius, ...) that are passed to the Deck class.}
}
\description{
Create a widget object using the Mapbox WebGL context
}
