% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeck.R
\name{rdeck}
\alias{rdeck}
\title{Create a widget object}
\usage{
rdeck(script, data, options = NULL, width = NULL, height = NULL,
  element_id = NULL, mapbox_properties = NULL, ...)
}
\arguments{
\item{script}{The name of the JavaScript file containing the vizualisation function.}

\item{data}{The data that is passed to the vizualisation function.}

\item{options}{(optional) Additional options that are passed to the vizualisation function.}

\item{width}{(optional) The width of the widget.}

\item{height}{(optional) The height of the widget.}

\item{element_id}{(optional) The unique id of the widget.}

\item{mapbox_properties}{(optional) The properties that are passed to the mapbox map object.
Only needed if the layers should be drawn into the same WebGL context that the Mapbox map is rendered in.
In this case you need to add your layers via \code{deck.MapboxLayer} to the \code{mapboxgl.Map} object.
This is useful if you want to combine Deck and Mapbox layers with each other in a smooth way.}

\item{...}{The properties (viewport settings, picking radius, ...) that are passed to the Deck class.}
}
\description{
Create a widget object
}
