% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_r2deck}
\alias{render_r2deck}
\title{Render r2deck using header def in JavaScript file}
\usage{
render_r2deck(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Render r2deck using header def in JavaScript file
}
