% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_r2deck_func}
\alias{render_r2deck_func}
\title{Render r2deck using header def in JavaScript file}
\usage{
render_r2deck_func(script = NULL)
}
\arguments{
\item{script}{The name of the JavaScript file containing the vizualisation function.}
}
\description{
Render r2deck using header def in JavaScript file
}
