% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{use_deps}
\alias{use_deps}
\title{Use dependencies}
\usage{
use_deps(dep_names)
}
\arguments{
\item{dep_names}{The names of the dependencies that should be used (see \code{\link{available_deps}}).}
}
\description{
Use dependencies
}
