% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapbox.R
\name{r2deck}
\alias{r2deck}
\title{Create a widget object}
\usage{
r2deck(script, data, longitude = -122.45, latitude = 37.8, zoom = 8,
  map_style = get_carto_style(), options = NULL, width = NULL,
  height = NULL, element_id = NULL, ...)
}
\arguments{
\item{script}{The name of the JavaScript file containing the vizualisation function.}

\item{data}{The data that is passed to the vizualisation function.}

\item{longitude}{...}

\item{latitude}{...}

\item{zoom}{...}

\item{map_style}{...}

\item{options}{(optional) Additional options that are passed to the vizualisation function as parameter \code{options}.}

\item{width}{(optional) The width of the widget.}

\item{height}{(optional) The height of the widget.}

\item{element_id}{(optional) The unique id of the widget.}

\item{...}{Additional properties (picking radius, ...) that are passed to the Deck class.}
}
\description{
Create a widget object
}
