% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeck.R
\name{rdeck}
\alias{rdeck}
\title{Create a widget object}
\usage{
rdeck(script, data, width = NULL, height = NULL, element_id = NULL,
  mapbox_properties = NULL, ...)
}
\arguments{
\item{script}{JavaScript file conatining the vizualisation function}

\item{data}{data passed to the vizualisation function in the JavaScript file}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{element_id}{unique id of the widget}

\item{mapbox_properties}{...}

\item{...}{...}
}
\description{
Create a widget object
}
